+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 RELEASE NOTES
 =============
 
+4.3
+---
+- Pause script on GUARD page
+- Track celestial coins as well
+- Add Summer Event profession
+- Add 'stop at lvl' opttion for each profession
+- Inactive characters are now greyed out
+
 4.2
 ---
 - Added "Resource Tracker" tab